+++ fuzzing/BUILD.bazel
@@ -1,16 +1,38 @@
-load("@rules_rust//rust:defs.bzl", "rust_binary")
+load("@rules_rust//rust:defs.bzl", "rust_binary", "rust_library")
+
+rust_library(
+    name = "fuzzing",
+    srcs = ["src/lib.rs"],
+    visibility = ["//fuzzing:__subpackages__"],
+    deps = [
+        "@crates//:anyhow",
+        "@crates//:thiserror",
+        "@crates//:tracing",
+        "@crates//:tracing-subscriber",
+    ],
+)
 
 rust_binary(
-    name = "fuzz",
-    srcs = ["main.rs"] + glob(["cmd/*.rs"]),
-    visibility = ["//visibility:public"],
+    name = "fuzz_test",
+    srcs = ["bin/fuzz_test.rs"],
+    visibility = ["//fuzzing:__subpackages__"],
     deps = [
+        ":fuzzing",
         "//process/entrypoint",
         "@crates//:anyhow",
         "@crates//:clap",
-        "@crates//:thiserror",
         "@crates//:tokio",
-        "@crates//:tracing",
-        "@crates//:tracing-subscriber",
+    ],
+)
+
+rust_binary(
+    name = "fuzz_corpus",
+    srcs = ["bin/fuzz_corpus.rs"],
+    visibility = ["//fuzzing:__subpackages__"],
+    deps = [
+        ":fuzzing",
+        "@crates//:anyhow",
+        "@crates//:clap",
+        "@crates//:tokio",
     ],
 )
