--- fuzzing/cmd/mod.rs
@@ -1,5 +1,0 @@
-mod corpus;
-mod test;
-
-pub use corpus::Corpus;
-pub use test::Test;
