+++ fuzzing/rust.bzl
@@ -1,5 +1,5 @@
 load("@rules_rust//rust:defs.bzl", "rust_binary")
-load("//fuzzing/private:fuzzing.bzl", "fuzz_corpus", "fuzz_test")
+load("//fuzzing/internal:fuzzing.bzl", "fuzz_corpus", "fuzz_test")
 
 def rust_fuzz_test(
         name,
@@ -10,8 +10,8 @@ def rust_fuzz_test(
     """A helper macro for fuzzing.
     """
 
-    target_name = name + "_target"
-    corpus_name = name + "_corpus"
+    test_name = name + "_fuzz_test"
+    corpus_name = name + "_fuzz_corpus"
 
     kwargs.setdefault("tags", []).extend([
         "fuzzing",
@@ -20,8 +20,8 @@ def rust_fuzz_test(
     fuzz_test(
         name = name,
         envs = envs,
+        executable = test_name,
         corpus = corpus_name,
-        target = target_name,
         tags = kwargs["tags"],
     )
 
@@ -43,7 +43,7 @@ def rust_fuzz_test(
     ])
 
     rust_binary(
-        name = target_name,
+        name = test_name,
         target_compatible_with = [
             "@rules_rust//rust/platform/channel:nightly",
         ],
